extracting RELEASE, using syncloud-id from sam dependencies as board identification

+++ syncloud.sh
@@ -10,7 +10,13 @@ fi
 set -x
 #export SHELLOPTS
 
-wget -qO- https://raw.githubusercontent.com/syncloud/apps/0.7/bootstrap.sh | bash
+RELEASE_FILE=/tmp/RELEASE
+if [ ! -f ${RELEASE_FILE} ]; then
+    echo "Relase file: ${RELEASE_FILE} is missing" 1>&2
+    exit 1
+fi
+
+wget -qO- https://raw.githubusercontent.com/syncloud/apps/$(<${RELEASE_FILE})/bootstrap.sh | bash
 
 sam install image-base
 sam install image-boot
